Remove ability to boolean check policies.
Policies can narrow down the subject scope,
so a boolean check does not carry enough information.

Instead, the caller should assert that the input and output subject
is the same, if they want to make sure the scope has not been narrowed.

The usefulness of policy scoping is that the policy can determine
how far to narrow down the scope, e.g.

```elixir
defmodule Blog do
   use Airframe.Policy

   def allow(Post, user, :list_posts) do
      {:ok, Airframe.scope(Post, user)}
   end

   def list_posts(user) do
      Post
      |> Airframe.allowed(Post, user)
      |> Repo.all()
      |> Airframe.tag(user)
   end
end
```